bugy #330 (su), #580 (hw tree), #695 (css), #687 (zakazane diskuse ke clankum) dale export pro newton it, ciselne url k diskusim ke clankum, zobrazit pocet clanku v rubrice, optimalizace cache sekci, podpora cache hw sekci, js stromecek pro hw sekci, poradna uz neni staticka
git-svn-id: https://zdrojaky.abclinuxu.cz/svn-abclinuxu/trunk@1862 be82d567-0701-4079-a524-1ea639e27e29

+++ src/cz/abclinuxu/servlets/utils/ServletUtils.java
@@ -194,7 +194,7 @@ public class ServletUtils implements Configurable {
                 ServletUtils.addError(PARAM_LOG_USER,"Přihlašovací jméno nenalezeno!",env, null);
                 return;
             }
-            user = new User(id.intValue());
+            user = new User(id);
             user = (User) persistence.findById(user);
 
             String password = (String) params.get(PARAM_LOG_PASSWORD);
